third_party/WebKit/LayoutTests/inspector/console/console-format-es6-expected.txt

CONSOLE MESSAGE: line 11: [object Object]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 11: Symbol()
CONSOLE MESSAGE: line 12: Symbol()
CONSOLE MESSAGE: line 11: Symbol(a)
CONSOLE MESSAGE: line 12: Symbol(a)
CONSOLE MESSAGE: line 11: [object Object]
CONSOLE MESSAGE: line 12: [object Object]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object WeakMap]
CONSOLE MESSAGE: line 12: [object WeakMap]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object WeakSet]
CONSOLE MESSAGE: line 12: [object WeakSet]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object Set]
CONSOLE MESSAGE: line 12: [object Set]
CONSOLE MESSAGE: line 11: [object Map]
CONSOLE MESSAGE: line 12: [object Map]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
CONSOLE MESSAGE: line 11: [object Map Iterator]
CONSOLE MESSAGE: line 12: [object Map Iterator]
CONSOLE MESSAGE: line 11: [object Set Iterator]
CONSOLE MESSAGE: line 12: [object Set Iterator]
Tests that console properly displays information about ES6 features.

console-format-es6.html:11 Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
console-format-es6.html:12 [Promise]
    0: Promise
    length: 1
    __proto__: Array[0]
globals[0]
Promise {[[PromiseStatus]]: "rejected", [[PromiseValue]]: -0}
console-format-es6.html:11 Symbol()
console-format-es6.html:12 [Symbol()]
    0: Symbol()
    length: 1
    __proto__: Array[0]
globals[1]
Symbol()
console-format-es6.html:11 Symbol(a)
console-format-es6.html:12 [Symbol(a)]
    0: Symbol(a)
    length: 1
    __proto__: Array[0]
globals[2]
Symbol(a)
console-format-es6.html:11 Object {getter: (...), a: Symbol(), Symbol(a): 2}
    a: Symbol()
    getter: (...)
    get getter: function getter() {}
    Symbol(a): 2
    __proto__: Object
console-format-es6.html:12 [Object]
    0: Object
    length: 1
    __proto__: Array[0]
globals[3]
Object {getter: (...), a: Symbol(), Symbol(a): 2}
    a: Symbol()
    getter: (...)
    get getter: function getter() {}
    Symbol(a): 2
    __proto__: Object
console-format-es6.html:11 Map {Object {getter: (...), a: Symbol(), Symbol(a): 2} => Object {foo: 1}}
    size: (...)
    __proto__: Map
    <entries>
console-format-es6.html:12 [Map]
    0: Map
    length: 1
    __proto__: Array[0]
globals[4]
Map {Object {getter: (...), a: Symbol(), Symbol(a): 2} => Object {foo: 1}}
    size: (...)
    __proto__: Map
    <entries>
console-format-es6.html:11 WeakMap {Object {getter: (...), a: Symbol(), Symbol(a): 2} => Object {foo: 1}}
    __proto__: WeakMap
    <entries>
console-format-es6.html:12 [WeakMap]
    0: WeakMap
    length: 1
    __proto__: Array[0]
globals[5]
WeakMap {Object {getter: (...), a: Symbol(), Symbol(a): 2} => Object {foo: 1}}
    __proto__: WeakMap
    <entries>
console-format-es6.html:11 Set {Object {getter: (...), a: Symbol(), Symbol(a): 2}}
    size: (...)
    __proto__: Set
    <entries>
console-format-es6.html:12 [Set]
    0: Set
    length: 1
    __proto__: Array[0]
globals[6]
Set {Object {getter: (...), a: Symbol(), Symbol(a): 2}}
    size: (...)
    __proto__: Set
    <entries>
console-format-es6.html:11 WeakSet {Object {getter: (...), a: Symbol(), Symbol(a): 2}}
    __proto__: WeakSet
    <entries>
console-format-es6.html:12 [WeakSet]
    0: WeakSet
    length: 1
    __proto__: Array[0]
globals[7]
WeakSet {Object {getter: (...), a: Symbol(), Symbol(a): 2}}
    __proto__: WeakSet
    <entries>
console-format-es6.html:11 Map {Map {} => WeakMap {}}
console-format-es6.html:12 [Map]
    0: Map
    length: 1
    __proto__: Array[0]
globals[8]
Map {Map {} => WeakMap {}}
console-format-es6.html:11 Map {Map {…} => WeakMap {…}}
    size: (...)
    __proto__: Map
    <entries>
console-format-es6.html:12 [Map]
    0: Map
    length: 1
    __proto__: Array[0]
globals[9]
Map {Map {…} => WeakMap {…}}
    size: (...)
    __proto__: Map
    <entries>
console-format-es6.html:11 Set {WeakSet {}}
console-format-es6.html:12 [Set]
    0: Set
    length: 1
    __proto__: Array[0]
globals[10]
Set {WeakSet {}}
console-format-es6.html:11 Set {WeakSet {…}}
    size: (...)
    __proto__: Set
    <entries>
console-format-es6.html:12 [Set]
    0: Set
    length: 1
    __proto__: Array[0]
globals[11]
Set {WeakSet {…}}
    size: (...)
    __proto__: Set
    <entries>
console-format-es6.html:11 Map {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, Object {foo: "from"} => Object {foo: "to"}}
console-format-es6.html:12 [Map]
    0: Map
    length: 1
    __proto__: Array[0]
globals[12]
Map {" from str " => " to str ", undefined => undefined, null => null, 42 => 42, Object {foo: "from"} => Object {foo: "to"}}
console-format-es6.html:11 MapIterator {41, Object {foo: 1}}
console-format-es6.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array[0]
globals[13]
MapIterator {41, Object {foo: 1}}
console-format-es6.html:11 MapIterator {42, Object {foo: 2}}
console-format-es6.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array[0]
globals[14]
MapIterator {42, Object {foo: 2}}
console-format-es6.html:11 MapIterator {[41, 42], [Object, Object]}
    __proto__: Object
    <entries>
console-format-es6.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array[0]
globals[15]
MapIterator {[41, 42], [Object, Object]}
    __proto__: Object
    <entries>
console-format-es6.html:11 SetIterator {41, Object {foo: 1}}
console-format-es6.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array[0]
globals[16]
SetIterator {41, Object {foo: 1}}
console-format-es6.html:11 SetIterator {41, Object {foo: 1}}
console-format-es6.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array[0]
globals[17]
SetIterator {41, Object {foo: 1}}
console-format-es6.html:11 SetIterator {[41, 41], [Object, Object]}
    __proto__: Object
    <entries>
console-format-es6.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array[0]
globals[18]
SetIterator {[41, 41], [Object, Object]}
    __proto__: Object
    <entries>
console-format-es6.html:11 MapIterator {Object {foo: 2}}
console-format-es6.html:12 [MapIterator]
    0: MapIterator
    length: 1
    __proto__: Array[0]
globals[19]
MapIterator {Object {foo: 2}}
console-format-es6.html:11 SetIterator {Object {foo: 1}}
console-format-es6.html:12 [SetIterator]
    0: SetIterator
    length: 1
    __proto__: Array[0]
globals[20]
SetIterator {Object {foo: 1}}
